removed Post-fix "Script" from Player and DemoProjectile
