+++ Assets/Scripts/DemoProjectile.cs
@@ -1,0 +1,21 @@
+
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class DemoProjectile : MonoBehaviour
+{
+	[SerializeField]
+	private float lifetime;
+
+	private void Awake()
+	{
+        Destroy(gameObject, lifetime);
+	}
+
+	private void OnCollisionEnter(Collision other)
+    {   
+		Destroy(gameObject);
+	}
+}
+
