--- Assets/Scripts/DemoProjectileScript.cs
@@ -1,21 +1,0 @@
-
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class DemoProjectileScript : MonoBehaviour
-{
-	[SerializeField]
-	private float lifetime;
-
-	private void Awake()
-	{
-        Destroy(gameObject, lifetime);
-	}
-
-	private void OnCollisionEnter(Collision other)
-    {   
-		Destroy(gameObject);
-	}
-}
-
